pyproject.toml
[project]
name = "itzn-api-py"
version = "0.1.0"
description = ""
authors = [
    {name = "Luaanny",email = "silva.luanny@escolar.ifrn.edu.br"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "google-api-python-client (>=2.187.0,<3.0.0)",
    "google-auth (>=2.43.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "itzn_api"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
run = 'fastapi dev app/main.py'
